+++ R/enrichDAVID.R
@@ -130,7 +130,7 @@ enrichDAVID <- function(gene,
 
     qobj <- tryCatch(qvalue(p=Over$pvalue, lambda=0.05, pi0.method="bootstrap"),
                      error=function(e) NULL)
-    if (class(qobj) == "qvalue") {
+    if (inherits(qobj, "qvalue")) {
         qvalues <- qobj$qvalues
     } else {
         qvalues <- NA