+++ R/utilities.R
@@ -1,3 +1,24 @@
+# Define a function to read the file with caching
+#' @importFrom fs path_join
+#' @importFrom digest digest
+#' @importFrom memoise memoise
+read_tsv_with_cache <- memoise::memoise(function(file_url,header = FALSE) {
+    cache_dir <- getOption("clusterProfiler_cache_dir")
+    # Generate a unique cache filename based on the file URL
+    cache_filename <- fs::path_join(c(cache_dir, paste0(digest::digest(file_url), ".rds")))
+    
+    # Check if the cached file exists
+    if (file.exists(cache_filename)) {
+        # If cached file exists, load and return the cached data
+        cached_data <- readRDS(cache_filename)
+        return(cached_data)
+    } else {
+        # If cached file does not exist, download and cache the data
+        data <- read.delim(url(file_url), sep = "\t", header = header)
+        saveRDS(data, cache_filename)
+        return(data)
+    }
+})
 
 
 mydownload <- function(url, method = NULL, quiet = TRUE, ...) {